Safari-609.1.11.1/LayoutTests/fast/events/touch/touch-event-pageXY-expected.txt

This tests pageX and pageY coordinates on touch events and touches.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS lastEvent.type is "touchstart"
PASS lastEvent.touches.length is 1
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 1
PASS lastEvent.touches[0].target.id is "touchtarget"
PASS lastEvent.touches[0].pageX is 8
PASS lastEvent.touches[0].pageY is 12
PASS lastEvent.touches[0].clientX is 8
PASS lastEvent.touches[0].clientY is 12
PASS lastEvent.touches[0].identifier is 0
PASS lastEvent.changedTouches[0].pageX is 8
PASS lastEvent.changedTouches[0].pageY is 12
PASS lastEvent.changedTouches[0].clientX is 8
PASS lastEvent.changedTouches[0].clientY is 12
PASS lastEvent.changedTouches[0].identifier is 0
PASS lastEvent.targetTouches[0].pageX is 8
PASS lastEvent.targetTouches[0].pageY is 12
PASS lastEvent.targetTouches[0].clientX is 8
PASS lastEvent.targetTouches[0].clientY is 12
PASS lastEvent.targetTouches[0].identifier is 0
PASS lastEvent.pageX is 8
PASS lastEvent.pageY is 12
move
PASS lastEvent.type is "touchmove"
PASS lastEvent.touches.length is 1
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 1
PASS lastEvent.touches[0].pageX is 10
PASS lastEvent.touches[0].pageY is 15
PASS lastEvent.touches[0].clientX is 10
PASS lastEvent.touches[0].clientY is 15
PASS lastEvent.touches[0].identifier is 0
PASS lastEvent.pageX is 10
PASS lastEvent.pageY is 15
add second touch
PASS lastEvent.type is "touchstart"
PASS lastEvent.touches.length is 2
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 2
PASS lastEvent.touches[0].pageX is 10
PASS lastEvent.touches[0].pageY is 15
PASS lastEvent.touches[0].clientX is 10
PASS lastEvent.touches[0].clientY is 15
PASS lastEvent.touches[0].identifier is 0
PASS lastEvent.touches[1].pageX is 40
PASS lastEvent.touches[1].pageY is 45
PASS lastEvent.touches[1].clientX is 40
PASS lastEvent.touches[1].clientY is 45
PASS lastEvent.touches[1].identifier is 1
PASS lastEvent.pageX is 25
PASS lastEvent.pageY is 30
end
PASS lastEvent.type is "touchend"
PASS lastEvent.touches.length is 1
PASS lastEvent.changedTouches.length is 1
PASS lastEvent.targetTouches.length is 1
PASS lastEvent.pageX is 40
PASS lastEvent.pageY is 45
PASS successfullyParsed is true

TEST COMPLETE
